+++ frontend/package.json
@@ -11,6 +11,7 @@
     "react-dom": "^17.0.2",
     "react-router-dom": "^5.2.1",
     "react-scripts": "5.0.0",
+    "sass": "^1.69.3",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
