+++ frontend/src/index.css
@@ -1,9 +1,6 @@
 @import url(./vendor/normalize.css);
 @import url(./vendor/fonts/fonts.css);
 
-/* Page Block */
-@import url(./blocks/page/page.css);
-
 /* Header Block */
 @import url(./blocks/header/header.css);
 @import url(./blocks/header/__nav/header__nav.css);
