+++ frontend/src/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
-import App from '../src/components/App';
+import App from '../src/components/App/App';
 import reportWebVitals from './reportWebVitals';
 import { HashRouter } from 'react-router-dom';
